+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.11.2 (unreleased)
+- Add benchmarks (using [Benchman](https://github.com/mar10/benchman)).
 
 ## 0.11.1 (2024-11-08)
 - `t0.diff(t1, ...)` adds nodes from t1 when possible, so the new status is